intermediate revision - updated sylink logic to exclusively collect the symlinks when flag is provided, switched to FLOAT(6) data type for allmetric fields, removed unneded initialization code

+++ datatxt-cachedb/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 -- 
 DROP TABLE IF EXISTS `metric_table`;
--- NOTE: BIGINT,TIMESTAMP, INTEGER, FLOAT are valid type for both SQLite and MySQL
-CREATE TABLE `metric_table` ( `id` BIGINT, `hostname` TEXT NOT NULL,  `timestamp` TIMESTAMP, `memory` INTEGER, `cpu` INTEGER, `disk` FLOAT(6), `load_average` INTEGER, PRIMARY KEY(`id`) );
+-- NOTE: BIGINT,TIMESTAMP, FLOAT(6), FLOAT are valid type for both SQLite and MySQL
+CREATE TABLE `metric_table` ( `id` BIGINT, `hostname` TEXT NOT NULL,  `timestamp` TIMESTAMP, `memory` FLOAT(6), `cpu` FLOAT(6), `disk` FLOAT(6), `load_average` FLOAT(6), PRIMARY KEY(`id`) );
